move docker client to provider/clients
