+++ core/provider/clients/docker_client.go
@@ -1,4 +1,4 @@
-package provider
+package clients
 
 import (
 	"context"
