+++ core/provider/docker/provider.go
@@ -36,7 +36,7 @@ type Provider struct {
 	state   *ProviderState
 	stateMu sync.Mutex
 
-	dockerClient           provider.DockerClient
+	dockerClient           clients.DockerClient
 	dockerNetworkAllocator *ipallocator.Range
 	networkMu              sync.Mutex
 	logger                 *zap.Logger
@@ -45,7 +45,7 @@ type Provider struct {
 var _ provider.ProviderI = (*Provider)(nil)
 
 func CreateProvider(ctx context.Context, logger *zap.Logger, providerName string) (*Provider, error) {
-	dockerClient, err := provider.NewDockerClient("")
+	dockerClient, err := clients.NewDockerClient("")
 	if err != nil {
 		return nil, err
 	}
@@ -118,7 +118,7 @@ func RestoreProvider(ctx context.Context, logger *zap.Logger, state []byte) (*Pr
 		logger: logger,
 	}
 
-	dockerClient, err := provider.NewDockerClient("")
+	dockerClient, err := clients.NewDockerClient("")
 	if err != nil {
 		return nil, err
 	}
